requirements.txt
selenium==4.23.1
pytest==8.3.3
pytest-html==4.1.1
python-dotenv
webdriver-manager
